Updated Version 5.0.3

+++ PKGBUILD
@@ -1,6 +1,6 @@
 # Maintainer: Jonathan Boyle <programmerfalcon@gmail.com>
 pkgname=mulle-bashfunctions
-pkgver=4.1.1
+pkgver=5.0.3
 pkgrel=1
 epoch=1
 pkgdesc="A collection of shell functions"
@@ -10,9 +10,9 @@ license=('BSD3')
 groups=('mulle-sde-all')
 depends=('bash')
 changelog="RELEASENOTES.md"
-source=("https://github.com/mulle-nat/mulle-bashfunctions/archive/4.1.1.tar.gz" "package.patch")
+source=("https://github.com/mulle-nat/mulle-bashfunctions/archive/5.0.3.tar.gz" "package.patch")
 noextract=()
-md5sums=('910b840579c70b831efb23d241879ddf' 'ae1135c609e963283f0840afd5d21f86')
+md5sums=('e1a8f6fbe947b4a2af966bb07fc2e5fa' '317ab45061adf8699fca047fabd90db3')
 validpgpkeys=()
 
 package() {
@@ -21,15 +21,16 @@ package() {
     install -Dm644 LICENSE "$pkgdir/usr/share/licenses/$pkgname/LICENSE"
 
     # Run Install
-    cd bin
-    chmod +x installer
-    ./installer "${pkgdir}/usr"
+    chmod +x bin/installer
+    bin/installer "${pkgdir}/usr"
 
     # Move libexec to lib/$pkgname
     cd "${pkgdir}/usr"
-    mkdir -p "lib/"
-    mv libexec/* "lib/"
-    rm -r libexec
+    if [ -d "libexec" ]; then
+        mkdir -p "lib/"
+        mv libexec/* "lib/"
+        rm -r libexec
+    fi
 }
 
 prepare() {
